Apply clang-tidy suggestions to all backends (#2839)
* Add clang-tidy configuration file

* Cleanup some exception code

* Add additional upstream directories to .gitignore

* Remove unused parameters from wrap and transform implementations

* Fix warnings and removed unused calls

+++ src/backend/cpu/math.cpp
@@ -16,7 +16,7 @@ uchar abs(uchar val) { return val; }
 uintl abs(uintl val) { return val; }
 
 cfloat scalar(float val) {
-    cfloat cval = {(float)val, 0};
+    cfloat cval = {val, 0};
     return cval;
 }
